src/score-maker/pyproject.toml
[tool.poetry]
name = "score-maker"
version = "0.1.0"
description = "Rating events by users"
authors = ["Nikita Pisarev <>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"
aiokafka = "^0.11.0"
uvicorn = "^0.31.0"
asgiref = "^3.8.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
